requirements.txt
streamlit==1.29.0
groq==0.4.1
sentence-transformers>=2.2.2,<3
transformers>=4.44.0,<5
PyPDF2==3.0.1
pymupdf==1.23.8
faiss-cpu==1.8.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.4
langdetect==1.0.9
langchain==0.0.350
httpx==0.27.0
torch==2.3.1
tokenizers>=0.21,<0.22
setuptools==69.0.3